requirements/base.txt
pytz==2020.1  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==8.0.0  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
pydash==4.8.0
freezegun==1.0.0
pendulum==2.1.2
bcrypt==3.2.0
cached-property==1.5.2

# Django
# ------------------------------------------------------------------------------
django==3.0.10  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-cron==0.5.1
django-filter==2.3.0
django-composite-foreignkey==1.1.0
pandas==1.1.5
requests-mock===1.7.0

# Django REST Framework
# ------------------------------------------------------------------------------
djangorestframework==3.11.1  # https://github.com/encode/django-rest-framework
djangorestframework-simplejwt==4.4.0
django-cors-headers==3.5.0 # https://github.com/adamchainz/django-cors-headers
drf-flex-fields==0.8.6
django-rest-passwordreset==1.1.0
drf-yasg==1.17.1
